raspetljavanje cirkularne zavisnosti servisa
