+++ berza/src/main/java/rs/edu/raf/banka/berza/service/impl/BerzaService.java
@@ -23,8 +23,6 @@ import java.util.*;
 @Service
 public class BerzaService {
 
-    private static Long berzaId = -1L;
-
     private BerzaRepository berzaRepository;
     private AkcijeRepository akcijeRepository;
     private OrderService orderService;
@@ -53,12 +51,6 @@ public class BerzaService {
 
     public List<Berza> findAll(){
         return berzaRepository.findAll();
-    }
-
-    public void addOrderToBerza(Order order, Long berzaId){
-        Berza berza = berzaRepository.findBerzaById(berzaId);
-        berza.getOrderi().add(order);
-        berzaRepository.save(berza);
     }
 
     public Berza findBerza(String oznaka){
@@ -115,7 +107,7 @@ public class BerzaService {
 
         if(hartijaId == -1L)
             return new OrderResponse("Error");
-        berzaId = berza;
+        OrderService.berzaId = berza;
 
         Double ukupnaCena = getPrice(ask, bid, orderAkcija);
         Double provizija = getCommission(ukupnaCena, orderType);
@@ -126,152 +118,9 @@ public class BerzaService {
                 provizija, orderType, isAON, isMargin, oznakaHartije, status, ask, bid);
 
         if(status == OrderStatus.APPROVED)
-            executeOrder(order.getId());
+            orderService.executeOrder(order.getId());
 
         return new OrderResponse(MessageUtils.ORDER_SUCCESSFUL);
-    }
-
-    @Async
-    public OrderResponse executeOrder(Long orderId) {
-        Order order = orderService.getOrder(orderId);
-
-        if(!order.getOrderStatus().equals(OrderStatus.APPROVED))
-            return new OrderResponse(MessageUtils.ORDER_REJECTED);
-
-        executeTransaction(berzaId, order, order.getAsk(), order.getBid());
-        order.setDone(true);
-        return new OrderResponse(MessageUtils.ORDER_SUCCESSFUL);
-    }
-
-    @Async
-    public OrderResponse executeTransaction(Long berzaId, Order order, Double ask, Double bid){
-        boolean flag = true;
-        if(berzaId != -1){
-            OrderStatusResponse orderStatus = this.getOrderStatus(berzaId);
-            flag = orderStatus.isBerzaOtvorena();
-        }
-        if(order.isAON()){
-            Transakcija transakcija;
-            if(flag)
-                transakcija = transactionOrder(order.getKolicina(), order, ask, bid);
-            else
-                transakcija = transactionOrderWithDelay(order.getKolicina(), order, ask, bid);
-            if(berzaId != -1)
-                this.addOrderToBerza(order, berzaId);
-            return new OrderResponse("OK");
-        }
-
-        return executeMiniTransactions(berzaId, order, flag, ask, bid);
-    }
-
-    /**
-     * Margin je povezan sa walletom korisnika koji ce biti detaljnije objasnjen u drugoj iteraciji
-     * s obzirom na to, bice obradjen nakon nastavka specifikacije
-     */
-    @Async
-    public OrderResponse executeMiniTransactions(Long berzaId, Order order, boolean flag, Double ask, Double bid){
-        Random random = new Random();
-        int kolicina = order.getKolicina();
-        int kolicinaZaTransakciju = random.nextInt(kolicina) + 1;
-
-        if(order.getOrderAction().equals(OrderAction.BUY) && !canExecuteTransactionBuy(order, bid))
-            return new OrderResponse("You can't proceed this action.");
-
-        if(order.getOrderAction().equals(OrderAction.SELL) && !canExecuteTransactionSell(order, ask))
-            return new OrderResponse("You can't proceed this action.");
-
-        while(kolicina - kolicinaZaTransakciju > 0){
-            //transakcija fixe delay
-            if(flag)
-                transactionOrder(kolicinaZaTransakciju, order, ask, bid);
-            else
-                transactionOrderWithDelay(kolicinaZaTransakciju, order, ask, bid);
-
-            kolicina -= kolicinaZaTransakciju;
-            kolicinaZaTransakciju = random.nextInt(kolicina) + 1;
-        }
-
-        orderService.finishOrder(order);
-
-        return new OrderResponse("OK");
-    }
-
-    @Async
-    Transakcija transactionOrder(Integer transactionAmount, Order order, Double ask, Double bid){
-        /**
-         * MARKET_ORDER se izvrsava odmah, pa nema potrebe da cekamo
-         */
-        if(!order.getOrderType().equals(OrderType.MARKET_ORDER)) {
-            try {
-                int s = new Random().nextInt(((int)(0.24*60/order.getKolicina() * 1000)));
-                Thread.sleep(s);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-        Transakcija transakcija = new Transakcija();
-        transakcija.setOrder(order);
-        transakcija.setCena(transactionAmount * order.getUkupnaCena());
-        transakcija.setKolicina(transactionAmount);
-        transakcija.setVremeTranskacije(new Date());
-        return transakcijaService.saveTranskacija(transakcija);
-    }
-
-    /**
-     * ukoliko je berza zatvorena prilikom ordera ili je u after-hours, korisnik ceka duze
-     */
-    @Async
-    public Transakcija transactionOrderWithDelay(Integer transactionAmount, Order order, Double ask, Double bid){
-        try {
-            //3s simulaciju 30 minuta
-            Thread.sleep(3000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-        return transactionOrder(transactionAmount, order, ask, bid);
-    }
-
-    public boolean canExecuteTransactionBuy(Order order, Double bid){
-        switch(order.getOrderType()){
-            case LIMIT_ORDER:
-                if(order.getUkupnaCena() <= order.getLimitValue())
-                    return true;
-                break;
-            case STOP_LIMIT_ORDER:
-                if(order.getUkupnaCena() <= order.getLimitValue() && order.getUkupnaCena() < bid){
-                    order.setOrderType(OrderType.LIMIT_ORDER);
-                    return true;
-                }
-                break;
-            case STOP_ORDER:
-                if(order.getUkupnaCena() < bid)
-                    return true;
-                break;
-            default:
-                return true;
-        }
-        return false;
-    }
-
-    public boolean canExecuteTransactionSell(Order order, Double ask){
-        switch(order.getOrderType()){
-            case LIMIT_ORDER:
-                if(order.getUkupnaCena() > order.getLimitValue())
-                    return true;
-                break;
-            case STOP_LIMIT_ORDER:
-                if(order.getUkupnaCena() > order.getLimitValue() && order.getUkupnaCena() > ask){
-                    order.setOrderType(OrderType.LIMIT_ORDER);
-                    return true;
-                }
-            case STOP_ORDER:
-                if(order.getUkupnaCena() > ask)
-                    return true;
-                break;
-            default:
-                return true;
-        }
-        return false;
     }
 
     public Double getPrice(Double ask, Double bid, OrderAction orderAction){
@@ -299,36 +148,6 @@ public class BerzaService {
         if(orderType.equals(OrderType.MARKET_ORDER))
             return Math.min(0.14 * price, 7);
         return Math.min(0.24 * price, 12);
-    }
-
-    public OrderStatusResponse getOrderStatus(Long id){
-        Berza berza = berzaRepository.findBerzaById(id);
-        Date date = new Date();
-
-        String openTime = berza.getOpenTime();
-        String closeTime = berza.getCloseTime();
-
-        DateFormat dateFormat = new SimpleDateFormat("hh:mm");
-        try {
-            if(isOverlapping(dateFormat.parse(openTime), dateFormat.parse(closeTime), date))
-                return new OrderStatusResponse(true, "Order odobren.");
-
-            if(differenceInHours(dateFormat.parse(closeTime), date))
-                return new OrderStatusResponse(false, "Berza je trenutno u after-hours stanju.");
-
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-        return new OrderStatusResponse(false, "Berza ne radi.");
-    }
-
-    private boolean isOverlapping(Date start, Date end, Date timeToCheck){
-        return start.before(timeToCheck) && end.after(timeToCheck);
-    }
-
-    public boolean differenceInHours(Date closingTime, Date timeToCheck){
-        long differenceInMilliSeconds = timeToCheck.getTime() - closingTime.getTime();
-        return (differenceInMilliSeconds / (60 * 60 * 1000)) % 24 <= 4;
     }
 
     private OrderType getOrderType(Integer limitValue, Integer stopValue) {
