+++ berza/src/main/java/rs/edu/raf/banka/berza/service/impl/OrderService.java
@@ -1,29 +1,44 @@
 package rs.edu.raf.banka.berza.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import rs.edu.raf.banka.berza.enums.*;
+import rs.edu.raf.banka.berza.model.Berza;
 import rs.edu.raf.banka.berza.model.Order;
+import rs.edu.raf.banka.berza.model.Transakcija;
+import rs.edu.raf.banka.berza.repository.BerzaRepository;
 import rs.edu.raf.banka.berza.repository.OrderRepository;
 import rs.edu.raf.banka.berza.response.ApproveRejectOrderResponse;
+import rs.edu.raf.banka.berza.response.OrderResponse;
+import rs.edu.raf.banka.berza.response.OrderStatusResponse;
 import rs.edu.raf.banka.berza.utils.MessageUtils;
 
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
+import java.util.Random;
 
 @Service
 public class OrderService {
 
+    public static Long berzaId;
+
     private OrderRepository orderRepository;
     private FuturesUgovoriPodaciService futuresUgovoriPodaciService;
-    private BerzaService berzaService;
+    private BerzaRepository berzaRepository;
+
+    private TransakcijaService transakcijaService;
 
     @Autowired
     public OrderService(OrderRepository orderRepository, FuturesUgovoriPodaciService futuresUgovoriPodaciService,
-                        BerzaService berzaService){
+                        TransakcijaService transakcijaService, BerzaRepository berzaRepository){
         this.orderRepository = orderRepository;
         this.futuresUgovoriPodaciService = futuresUgovoriPodaciService;
-        this.berzaService = berzaService;
+        this.transakcijaService = transakcijaService;
+        this.berzaRepository = berzaRepository;
     }
 
     public Order getOrder(Long id) {
@@ -53,7 +68,7 @@ public class OrderService {
         order.setLastModified(new Date());
         orderRepository.save(order);
 
-        berzaService.executeOrder(id);
+        executeOrder(id);
 
         return new ApproveRejectOrderResponse(MessageUtils.ORDER_APPROVED);
     }
@@ -99,6 +114,185 @@ public class OrderService {
         order.setDone(true);
         order.setLastModified(new Date());
         orderRepository.save(order);
+    }
+
+    @Async
+    public OrderResponse executeOrder(Long orderId) {
+        Order order = getOrder(orderId);
+
+        if(!order.getOrderStatus().equals(OrderStatus.APPROVED))
+            return new OrderResponse(MessageUtils.ORDER_REJECTED);
+
+        executeTransaction(berzaId, order, order.getAsk(), order.getBid());
+        order.setDone(true);
+        return new OrderResponse(MessageUtils.ORDER_SUCCESSFUL);
+    }
+
+    @Async
+    public OrderResponse executeTransaction(Long berzaId, Order order, Double ask, Double bid){
+        boolean flag = true;
+        if(berzaId != -1){
+            OrderStatusResponse orderStatus = getOrderStatus(berzaId);
+            flag = orderStatus.isBerzaOtvorena();
+        }
+        if(order.isAON()){
+            Transakcija transakcija;
+            if(flag)
+                transakcija = transactionOrder(order.getKolicina(), order, ask, bid);
+            else
+                transakcija = transactionOrderWithDelay(order.getKolicina(), order, ask, bid);
+            if(berzaId != -1)
+                addOrderToBerza(order, berzaId);
+            return new OrderResponse("OK");
+        }
+
+        return executeMiniTransactions(berzaId, order, flag, ask, bid);
+    }
+
+    /**
+     * Margin je povezan sa walletom korisnika koji ce biti detaljnije objasnjen u drugoj iteraciji
+     * s obzirom na to, bice obradjen nakon nastavka specifikacije
+     */
+    @Async
+    public OrderResponse executeMiniTransactions(Long berzaId, Order order, boolean flag, Double ask, Double bid){
+        Random random = new Random();
+        int kolicina = order.getKolicina();
+        int kolicinaZaTransakciju = random.nextInt(kolicina) + 1;
+
+        if(order.getOrderAction().equals(OrderAction.BUY) && !canExecuteTransactionBuy(order, bid))
+            return new OrderResponse("You can't proceed this action.");
+
+        if(order.getOrderAction().equals(OrderAction.SELL) && !canExecuteTransactionSell(order, ask))
+            return new OrderResponse("You can't proceed this action.");
+
+        while(kolicina - kolicinaZaTransakciju > 0){
+            //transakcija fixe delay
+            if(flag)
+                transactionOrder(kolicinaZaTransakciju, order, ask, bid);
+            else
+                transactionOrderWithDelay(kolicinaZaTransakciju, order, ask, bid);
+
+            kolicina -= kolicinaZaTransakciju;
+            kolicinaZaTransakciju = random.nextInt(kolicina) + 1;
+        }
+
+        finishOrder(order);
+
+        return new OrderResponse("OK");
+    }
+
+    @Async
+    Transakcija transactionOrder(Integer transactionAmount, Order order, Double ask, Double bid){
+        /**
+         * MARKET_ORDER se izvrsava odmah, pa nema potrebe da cekamo
+         */
+        if(!order.getOrderType().equals(OrderType.MARKET_ORDER)) {
+            try {
+                int s = new Random().nextInt(((int)(0.24*60/order.getKolicina() * 1000)));
+                Thread.sleep(s);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+        }
+        Transakcija transakcija = new Transakcija();
+        transakcija.setOrder(order);
+        transakcija.setCena(transactionAmount * order.getUkupnaCena());
+        transakcija.setKolicina(transactionAmount);
+        transakcija.setVremeTranskacije(new Date());
+        return transakcijaService.saveTranskacija(transakcija);
+    }
+
+    /**
+     * ukoliko je berza zatvorena prilikom ordera ili je u after-hours, korisnik ceka duze
+     */
+    @Async
+    public Transakcija transactionOrderWithDelay(Integer transactionAmount, Order order, Double ask, Double bid){
+        try {
+            //3s simulaciju 30 minuta
+            Thread.sleep(3000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+        return transactionOrder(transactionAmount, order, ask, bid);
+    }
+
+    public void addOrderToBerza(Order order, Long berzaId){
+        Berza berza = berzaRepository.findBerzaById(berzaId);
+        berza.getOrderi().add(order);
+        berzaRepository.save(berza);
+    }
+
+    public boolean canExecuteTransactionBuy(Order order, Double bid){
+        switch(order.getOrderType()){
+            case LIMIT_ORDER:
+                if(order.getUkupnaCena() <= order.getLimitValue())
+                    return true;
+                break;
+            case STOP_LIMIT_ORDER:
+                if(order.getUkupnaCena() <= order.getLimitValue() && order.getUkupnaCena() < bid){
+                    order.setOrderType(OrderType.LIMIT_ORDER);
+                    return true;
+                }
+                break;
+            case STOP_ORDER:
+                if(order.getUkupnaCena() < bid)
+                    return true;
+                break;
+            default:
+                return true;
+        }
+        return false;
+    }
+
+    public boolean canExecuteTransactionSell(Order order, Double ask){
+        switch(order.getOrderType()){
+            case LIMIT_ORDER:
+                if(order.getUkupnaCena() > order.getLimitValue())
+                    return true;
+                break;
+            case STOP_LIMIT_ORDER:
+                if(order.getUkupnaCena() > order.getLimitValue() && order.getUkupnaCena() > ask){
+                    order.setOrderType(OrderType.LIMIT_ORDER);
+                    return true;
+                }
+            case STOP_ORDER:
+                if(order.getUkupnaCena() > ask)
+                    return true;
+                break;
+            default:
+                return true;
+        }
+        return false;
+    }
+
+    public OrderStatusResponse getOrderStatus(Long id){
+        Berza berza = berzaRepository.findBerzaById(id);
+        Date date = new Date();
+
+        String openTime = berza.getOpenTime();
+        String closeTime = berza.getCloseTime();
+
+        DateFormat dateFormat = new SimpleDateFormat("hh:mm");
+        try {
+            if(isOverlapping(dateFormat.parse(openTime), dateFormat.parse(closeTime), date))
+                return new OrderStatusResponse(true, MessageUtils.ORDER_APPROVED);
+
+            if(differenceInHours(dateFormat.parse(closeTime), date))
+                return new OrderStatusResponse(false, "Berza je trenutno u after-hours stanju.");
+
+        } catch (ParseException e) {
+            e.printStackTrace();
+        }
+        return new OrderStatusResponse(false, "Berza ne radi.");
+    }
+
+    private boolean isOverlapping(Date start, Date end, Date timeToCheck){
+        return start.before(timeToCheck) && end.after(timeToCheck);
+    }
+
+    public boolean differenceInHours(Date closingTime, Date timeToCheck){
+        long differenceInMilliSeconds = timeToCheck.getTime() - closingTime.getTime();
+        return (differenceInMilliSeconds / (60 * 60 * 1000)) % 24 <= 4;
     }
 
     private boolean isSupervisor(String userRole) {
